change price to BigDecimal

+++ src/main/java/pl/bykowski/week2/Product.java
@@ -2,13 +2,14 @@ package pl.bykowski.week2;
 
 import lombok.Data;
 import org.apache.commons.lang3.StringUtils;
+
+import java.math.BigDecimal;
 import java.util.Objects;
 
 @Data
 public class Product {
     private final String name;
-    // BigDecimal instead of Integer
-    private final Integer price;
+    private final BigDecimal price;
 
     public boolean isValidated() {
         return StringUtils.isNotEmpty(name) && Objects.nonNull(price);
